Handle payment gateways callbacks
