+++ src/@next/pages/CheckoutPage/subpages/CheckoutPaymentSubpage.tsx
@@ -11,16 +11,16 @@ import { RouteComponentProps } from "react-router";
 import { CheckoutPayment } from "@components/organisms";
 import { useCheckout } from "@saleor/sdk";
 import { commonMessages } from "@temp/intl";
-import { ICardData, IFormError } from "@types";
+import { IFormError } from "@types";
 
 export interface ICheckoutPaymentSubpageHandles {
   submitPayment: () => void;
 }
 interface IProps extends RouteComponentProps<any> {
-  paymentGatewayFormRef?: React.RefObject<HTMLFormElement>;
-  selectPaymentGateway: (paymentGateway: string) => void;
+  paymentGatewayFormRef: React.RefObject<HTMLFormElement>;
   changeSubmitProgress: (submitInProgress: boolean) => void;
   onSubmitSuccess: () => void;
+  onPaymentGatewayError: (errors: IFormError[]) => void;
 }
 
 const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
@@ -30,24 +30,15 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
   {
     paymentGatewayFormRef,
     changeSubmitProgress,
-    selectPaymentGateway,
     onSubmitSuccess,
+    onPaymentGatewayError,
     ...props
   }: IProps,
   ref
 ) => {
-  const {
-    availablePaymentGateways,
-    promoCodeDiscount,
-    addPromoCode,
-    removePromoCode,
-    createPayment,
-  } = useCheckout();
+  const { promoCodeDiscount, addPromoCode, removePromoCode } = useCheckout();
 
-  const [gatewayErrors, setGatewayErrors] = useState<IFormError[]>([]);
   const [promoCodeErrors, setPromoCodeErrors] = useState<IFormError[]>([]);
-
-  const paymentGateways = availablePaymentGateways || [];
 
   const promoCodeDiscountFormId = "discount-form";
   const promoCodeDiscountFormRef = useRef<HTMLFormElement>(null);
@@ -59,37 +50,19 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
         promoCodeDiscountFormRef.current?.dispatchEvent(
           new Event("submit", { cancelable: true })
         );
-      } else if (paymentGatewayFormRef?.current) {
+      } else if (paymentGatewayFormRef.current) {
         paymentGatewayFormRef.current.dispatchEvent(
           new Event("submit", { cancelable: true })
         );
       } else {
         changeSubmitProgress(false);
-        setGatewayErrors([
+        onPaymentGatewayError([
           { message: intl.formatMessage(commonMessages.choosePaymentMethod) },
         ]);
       }
     },
   }));
 
-  const handleProcessPayment = async (
-    gateway: string,
-    token: string,
-    cardData?: ICardData
-  ) => {
-    const { dataError } = await createPayment(gateway, token, cardData);
-    const errors = dataError?.error;
-    changeSubmitProgress(false);
-    if (errors) {
-      setGatewayErrors(errors);
-    } else {
-      setGatewayErrors([]);
-      onSubmitSuccess();
-    }
-  };
-  const handlePaymentGatewayError = () => {
-    changeSubmitProgress(false);
-  };
   const handleAddPromoCode = async (promoCode: string) => {
     const { dataError } = await addPromoCode(promoCode);
     const errors = dataError?.error;
@@ -98,13 +71,13 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
       setPromoCodeErrors(errors);
     } else {
       setPromoCodeErrors([]);
-      if (paymentGatewayFormRef?.current) {
+      if (paymentGatewayFormRef.current) {
         paymentGatewayFormRef.current.dispatchEvent(
           new Event("submit", { cancelable: true })
         );
       } else {
         changeSubmitProgress(false);
-        setGatewayErrors([
+        onPaymentGatewayError([
           { message: intl.formatMessage(commonMessages.choosePaymentMethod) },
         ]);
       }
@@ -118,26 +91,26 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
       setPromoCodeErrors(errors);
     } else {
       setPromoCodeErrors([]);
-      if (paymentGatewayFormRef?.current) {
+      if (paymentGatewayFormRef.current) {
         paymentGatewayFormRef.current.dispatchEvent(
           new Event("submit", { cancelable: true })
         );
       } else {
         changeSubmitProgress(false);
-        setGatewayErrors([
+        onPaymentGatewayError([
           { message: intl.formatMessage(commonMessages.choosePaymentMethod) },
         ]);
       }
     }
   };
   const handleSubmitUnchangedDiscount = () => {
-    if (paymentGatewayFormRef?.current) {
+    if (paymentGatewayFormRef.current) {
       paymentGatewayFormRef.current.dispatchEvent(
         new Event("submit", { cancelable: true })
       );
     } else {
       changeSubmitProgress(false);
-      setGatewayErrors([
+      onPaymentGatewayError([
         { message: intl.formatMessage(commonMessages.choosePaymentMethod) },
       ]);
     }
@@ -146,9 +119,6 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
   return (
     <CheckoutPayment
       {...props}
-      gatewayErrors={gatewayErrors}
-      paymentGateways={paymentGateways}
-      selectPaymentGateway={selectPaymentGateway}
       promoCodeDiscountFormId={promoCodeDiscountFormId}
       promoCodeDiscountFormRef={promoCodeDiscountFormRef}
       promoCodeDiscount={{
@@ -158,8 +128,6 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
       removeVoucherCode={handleRemovePromoCode}
       submitUnchangedDiscount={handleSubmitUnchangedDiscount}
       promoCodeErrors={promoCodeErrors}
-      processPayment={handleProcessPayment}
-      onGatewayError={handlePaymentGatewayError}
     />
   );
 };
